New lib updates
