+++ sims.code-workspace
@@ -176,6 +176,7 @@
       "Formio",
       "golevelup",
       "Ilike",
+      "jsonata",
       "lastupdated",
       "LGCY",
       "LGCYAAAA",
@@ -189,6 +190,7 @@
       "NSLSC",
       "Overaward",
       "overawards",
+      "partnerinformationandincomeappeal",
       "PDAP",
       "PDRJ",
       "PDRQ",
@@ -199,11 +201,16 @@
       "PSFS",
       "PTSSR",
       "PTWTHD",
+      "roomandboardcostsappeal",
       "SABC",
       "sbsd",
       "SFAS",
       "siteprotected",
       "SSRN",
+      "studentadditionaltransportationappeal",
+      "studentdependantsappeal",
+      "studentdisabilityappeal",
+      "studentfinancialinformationappeal",
       "supplieraddress",
       "suppliername",
       "suppliernumber",
@@ -213,13 +220,7 @@
       "typeorm",
       "unparse",
       "WTHD",
-      "Zeebe",
-      "studentdependantsappeal",
-      "studentadditionaltransportationappeal",
-      "studentdisabilityappeal",
-      "studentfinancialinformationappeal",
-      "partnerinformationandincomeappeal",
-      "roomandboardcostsappeal"
+      "Zeebe"
     ],
     "[json]": {
       "editor.defaultFormatter": "vscode.json-language-features"
