+++ sources/packages/backend/package.json
@@ -79,7 +79,7 @@
     "express-basic-auth": "^1.2.1",
     "helmet": "^7.1.0",
     "ioredis": "^5.4.1",
-    "jsonpath": "^1.1.1",
+    "jsonata": "^2.1.0",
     "jwt-decode": "^3.1.2",
     "papaparse": "^5.4.1",
     "passport": "^0.6.0",
